Revert "Addressing @phinze's comments from https://github.com/hashicorp/terraform/pull/4401"

+++ builtin/providers/google/resource_pubsub_subscription.go
@@ -75,7 +75,7 @@ func resourcePubsubSubscriptionCreate(d *schema.ResourceData, meta interface{}) 
 	var ackDeadlineSeconds int64
 	ackDeadlineSeconds = 10
 	if v, ok := d.GetOk("ack_deadline_seconds"); ok {
-		ackDeadlineSeconds = int64(v.(int))
+		ackDeadlineSeconds = v.(int64)
 	}
 
 	var subscription *pubsub.Subscription
